update to alpha beta
alpha beta is sort of working. Theres two known bugs: for some reason my getNextMoves function will not always give correct middle moves (it will not pass through the if(brd.board[curRow+dir][curCol] != opp) statement because it thinks theres an opponent there. 
Also, It is still returning a 0,0,0,0 move, which i need to look deeper into. Currently its in puzzle mode and will print a large, confusing amount of stuff (i was using it to debug but will be confusing to others most likely). The issue right now with alpha beta has something to do with when it hits a terminal node; i dont know if my algorithm is set up properly as far as dealing with the terminal nodes. thats worth looking into, because it will show a -999 score as the max score, which makes no sense.